+++ src/z80asm/t/errors.t
@@ -558,7 +558,7 @@ t_z80asm_capture("-r0 -b -i".lib_file()." ".asm_file(), "",
 #------------------------------------------------------------------------------
 unlink_testfiles();
 
-my $objs = "errors.o error_func.o scan.o lib/array.o lib/class.o lib/str.o lib/strhash.o lib/list.o lib/fileutil.o options.o model.o module.o sym.o symtab.o symref.o codearea.o expr.o listfile.o lib/srcfile.o hist.o lib/dbg.o";
+my $objs = "errors.o error_func.o scan.o lib/array.o lib/class.o lib/str.o lib/strhash.o lib/list.o lib/fileutil.o options.o model.o module.o sym.o symtab.o codearea.o expr.o listfile.o lib/srcfile.o hist.o lib/dbg.o";
 
 # get init code except init() and main()
 my $init = <<'END';