+++ src/z80asm/t/objfile.t
@@ -4455,7 +4455,7 @@ t_binary(read_binfile(bin_file()), "\xC3\x00\x00");
 #------------------------------------------------------------------------------
 unlink_testfiles();
 
-my $objs = "objfile.o lib/class.o lib/array.o errors.o error_func.o lib/str.o lib/strhash.o lib/list.o lib/fileutil.o scan.o options.o model.o module.o sym.o symtab.o symref.o lib/srcfile.o hist.o expr.o listfile.o codearea.o lib/dbg.o";
+my $objs = "objfile.o lib/class.o lib/array.o errors.o error_func.o lib/str.o lib/strhash.o lib/list.o lib/fileutil.o scan.o options.o model.o module.o sym.o symtab.o lib/srcfile.o hist.o expr.o listfile.o codearea.o lib/dbg.o";
 
 # get init code except init() and main()
 my $init = <<'END';